requirements.txt
numpy==1.23.5
ollama==0.3.3
openai-whisper==20231117
pyneuphonic==1.1.3
sounddevice==0.4.7
torch==2.3.1